[FEAT] 하드코딩 값 바인딩 및 버그 수정

+++ src/main/java/susussg/pengreenlive/user/Mapper/UserMapper.java
@@ -9,7 +9,7 @@ import susussg.pengreenlive.user.dto.UpdateUserFormDTO;
 public interface UserMapper {
 
     String selectUserNmByUUID(String uuid);
-    String selectChannelNmByVendorSeq(int vendorSeq);
+    String selectChannelNmByVendorSeq(Long vendorSeq);
 
     void insertUser(SignupFormDTO signupForm);
